+++ dphonebook/lib/numberprovider.py
@@ -80,7 +80,7 @@ class NumberProvider(Thread):
         if fuzzy_time == 'one minute ago':
             return datetime.datetime.now() - datetime.timedelta(minutes=1)
 
-        time_components = re.search(r'(\d{1,2}) (sec|second|seconds|min|minute|hour|h|day|month)s? ago', fuzzy_time)
+        time_components = re.search(r'(\d{1,2}) ?(sec|second|seconds|min|minute|hour|h|day|month)s? ago', fuzzy_time)
         if not time_components:
             return None
         time_quantity = int(time_components.group(1))  # 12