+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `sms-receive.net` provider
 - Add `7sim.net` provider
+- Add `freereceivesms.com` provider
 
 ## [0.4.1] - 2022-06-24